+++ lib/plugin.js
@@ -3,6 +3,7 @@
 const _ = require('lodash')
 const routes = require('./routes')()
 const adapterUtils = require('./utils/adapter')()
+const routeUtils = require('./utils/route')()
 
 exports.register = function (server, opts, next) {
     server.expose('version', require('../package.json').version);
@@ -16,33 +17,8 @@ exports.register = function (server, opts, next) {
         next()
     });
     
-    const createOptionsRoute = function(schema) {
-        const tables = _.map(server.table()[0].table)
-        
-        //see if the options method already exists, if so, don't duplicate it
-        if (_.find(tables, {path : '/' + schema.type, method: 'options'})) return;
-        
-        server.route(_.merge(routes.options(schema), {
-            handler: (req, reply) => {
-                const tables = _.map(req.server.table()[0].table, (table) => {
-                    return _.pick(table, 'path', 'method')
-                })
-                
-                const pathVerbs = _.chain(tables)
-                .filter((table) => {
-                    return table.path.replace('/{id}', '') === req.path 
-                })
-                .pluck('method')
-                .map((verb) => { return verb.toUpperCase() })
-                .value();
-                
-                reply().header('Allow', pathVerbs.join(','))
-            }
-        }))
-    }
-    
     const get = function (schema) {
-        createOptionsRoute(schema)
+        routeUtils.createOptionsRoute(server, schema)
         adapter.processSchema(schema)
         return _.merge(routes.get(schema), {
             handler: (req, reply) => {
@@ -52,7 +28,7 @@ exports.register = function (server, opts, next) {
     }
     
     const getById = function (schema) {
-        createOptionsRoute(schema)
+        routeUtils.createOptionsRoute(server, schema)
         adapter.processSchema(schema)
         return _.merge(routes.getById(schema), {
             handler: (req, reply) => {
@@ -62,7 +38,7 @@ exports.register = function (server, opts, next) {
     }
     
     const post = function (schema) {
-        createOptionsRoute(schema)
+        routeUtils.createOptionsRoute(server, schema)
         adapter.processSchema(schema)
         return _.merge(routes.post(schema), {
             handler: (req, reply) => {
@@ -72,7 +48,7 @@ exports.register = function (server, opts, next) {
     }
     
     const patch = function (schema) {
-        createOptionsRoute(schema)
+        routeUtils.createOptionsRoute(server, schema)
         adapter.processSchema(schema)
         return _.merge(routes.patch(schema), {
             handler: (req, reply) => {
@@ -82,7 +58,7 @@ exports.register = function (server, opts, next) {
     }
     
     const del = function (schema) {
-        createOptionsRoute(schema)
+        routeUtils.createOptionsRoute(server, schema)
         adapter.processSchema(schema)
         return _.merge(routes.delete(schema), {
             handler: (req, reply) => {
